feat(product): 品牌管理实现 + 校验分组/错误码基础设施
按 brand-management.md 设计文档落地品牌管理 7 个接口：
分页查询/详情/新增/修改/状态更新/删除/分类下品牌。

实现：
- BrandController/Service（含 @Transactional 双表事务、乐观锁、引用检查）
- DTO/VO 5 个（BrandSaveDTO 用 Create/Update 校验分组区分增改）
- 关联表同步用 diff 规避逻辑删除与唯一约束冲突
- Brand.showStatus Byte→Integer 对齐规范
- product 补 MyBatisConfig(@MapperScan)，修复 Mapper 未扫描缺失

mall-common 基础设施：
- 新增校验分组 Create/Update 标记接口（规范要求，项目此前缺失）
- ResultCode 补品牌错误码 53001-53005

测试（mvn test 58 通过）：
- BrandServiceTest 纯单元测试（17）+ BrandControllerTest 切片测试（14）
- 修复 LambdaUpdateWrapper.set() 纯单测 lambda 缓存未初始化问题
  （@BeforeAll TableInfoHelper.initTableInfo 预热）

文档：
- brand-management.md 设计文档 + product-brand-demo.http
- docs/mall-product/PROGRESS.md 服务进度
- CLAUDE.md 补 JDK21 路径约定

+++ mall-product/src/main/java/com/mymall/product/entity/Brand.java
@@ -33,7 +33,7 @@ public class Brand extends BaseEntity {
     /**
      * 显示状态[0-不显示；1-显示]
      */
-    private Byte showStatus;
+    private Integer showStatus;
 
     /**
      * 检索首字母
